+++ app/Http/Controllers/UserStoryController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\UserStoryCreateRequest;
+use App\Http\Requests\UserStoryUpdateRequest;
 use App\Models\UserStory;
 use Illuminate\Http\Request;
 
@@ -23,25 +25,56 @@ class UserStoryController extends Controller
         // Logic to get a specific user story by ID
     }
 
-    public function store(Request $request)
+    public function store(UserStoryCreateRequest $request, $organizationId, $projectId)
     {
         // Logic to create a new user story
+        if ($request->user()->cannot('create', UserStory::class)) {
+            abort(403, 'Unauthorized action.');
+        }
+
+        $validated = $request->validated();
+
+        $user_story = UserStory::create([
+            'title' => $validated['title'],
+            'description' => $validated['description'],
+            'due_date' => $validated['due_date'],
+            'priority_id' => $validated['priority_id'],
+            'project_id' => $projectId,
+        ]);
+
         return response()->json([
             'message' => 'User story created successfully',
-            'user_story' => [
-                'id' => 1,
-                'title' => $request->input('title'),
-            ]
+            'user_story' => $user_story,
         ], 201);
     }
 
-    public function update(Request $request, $id)
+    public function update(UserStoryUpdateRequest $request, $organizationId, $projcetId, $userStoryId)
     {
         // Logic to update an existing user story
+
+        $validated = $request->validated();
+
+        if ($request->user()->cannot('update', UserStory::class, $userStoryId)) {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
+        $userStory = UserStory::findOrFail($userStoryId);
+
+        $userStory->fill($validated);
+        $userStory->save();
+
+        return response()->json([
+            'message' => 'User story updated successfully',
+            'user_story' => $userStory,
+        ], 200);
+
     }
 
-    public function destroy($organizationId, $projectId, $userStoryId)
+    public function destroy(Request $request, $organizationId, $projectId, $userStoryId)
     {
+        if ($request->user()->cannot('delete', UserStory::class, $projectId)) {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
+
         $userStory = UserStory::findOrFail($userStoryId);
 
         $userStory->delete();
